Create README.md
🦠 COVID-19 Detection: A Deep Learning-Powered Web App 📸
I have undertaken a project aimed at creating a machine-learning model capable of accurately detecting whether a patient has tested positive for COVID-19. In addition to this, I have developed a website utilizing Flask as the web framework, with the backend seamlessly integrating deep learning libraries and modules such as TensorFlow and Keras. The purpose of this website is to enable users to upload their X-ray or CT-scan images for the purpose of determining whether the model classifies them as carriers of COVID-19 or not.